requirements.txt
# Core
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0

# Geospatial
rasterio>=1.2.0
geopandas>=0.10.0
shapely>=1.8.0
pyproj>=3.0.0
fiona>=1.8.0

# Deep Learning
torch>=1.10.0
torchvision>=0.11.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.0.0
timm>=0.5.0

# Training utilities
tensorboard>=2.7.0
wandb>=0.12.0
tqdm>=4.62.0
pyyaml>=5.4.0
omegaconf>=2.1.0

# Evaluation
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Demo apps
streamlit>=1.0.0
fastapi>=0.70.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Utilities
click>=8.0.0
rich>=10.0.0
